% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_wordstream_data.R
\name{prepare_wordstream_data}
\alias{prepare_wordstream_data}
\title{Prepara os dados para a visualização WordStream}
\usage{
prepare_wordstream_data(data, time_col, group_col, word_col, top_n = 60)
}
\arguments{
\item{data}{Um data frame.}

\item{time_col}{O nome da coluna (string) que marca o tempo (ex: "Week", "Year").}

\item{group_col}{O nome da coluna (string) que define a categoria/grupo (ex: "POS", "Topic").}

\item{word_col}{O nome da coluna (string) que contém as palavras ou termos.}

\item{top_n}{O número de palavras mais frequentes a serem mantidas para cada grupo em cada período.}
}
\value{
Uma lista de listas no formato adequado para a função \code{wordstream}.
}
\description{
Esta função recebe um data frame em formato "tidy" (um registro por palavra),
calcula as frequências e o transforma na estrutura de lista aninhada
requerida pela função \code{wordstream()}.
}
