% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordstream.R
\name{renderWordstream}
\alias{renderWordstream}
\title{Render a wordstream in Shiny}
\usage{
renderWordstream(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a wordstream}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Render a wordstream in Shiny
}
