% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordstream.R
\name{wordstream}
\alias{wordstream}
\title{Cria uma visualização WordStream a partir de dados pré-formatados}
\usage{
wordstream(
  data,
  width = NULL,
  height = NULL,
  min_font_size = 14,
  max_font_size = 30,
  top_words_vis = 60
)
}
\arguments{
\item{data}{Uma lista de listas pré-formatada, geralmente criada pela
função \code{prepare_wordstream_data}. A estrutura deve ser:
\code{list(list(date="Week 1", words=list(NOUN=list(...), ADJ=list(...))), ...)}}

\item{width}{Largura do widget em pixels.}

\item{height}{Altura do widget em pixels.}

\item{min_font_size}{Tamanho mínimo da fonte.}

\item{max_font_size}{Tamanho máximo da fonte.}

\item{top_words_vis}{Número de palavras a serem renderizadas pela biblioteca JS.
Note que a filtragem principal é feita em \code{prepare_wordstream_data}.}
}
\description{
Cria uma visualização WordStream a partir de dados pré-formatados
}
