% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordstream.R
\name{wordstreamOutput}
\alias{wordstreamOutput}
\title{Shiny binding for wordstream}
\usage{
wordstreamOutput(outputId, width = "100\%", height = "500px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit
(like \code{'100\%'}, \code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}
}
\description{
Shiny binding for wordstream
}
